feat(#1): change entity files' name
