+++ src/app.module.ts
@@ -2,7 +2,6 @@ import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { MariaDbConfigProvider } from './config/typeorm.config';
 import { ConfigModule } from '@nestjs/config';
 import { AdminOptionModule } from './admin-option/admin-option.module';
 import { ApplicationModule } from './application/application.module';
@@ -23,6 +22,7 @@ import { RootModule } from './root/root.module';
 import { S3Module } from './s3/s3.module';
 import { ScheduleModule } from './schedule/schedule.module';
 import { SearchModule } from './search/search.module';
+import { typeORMConfig } from './config/typeorm.config';
 
 @Module({
   imports: [
@@ -30,9 +30,7 @@ import { SearchModule } from './search/search.module';
       cache: true,
       isGlobal: true,
     }),
-    TypeOrmModule.forRootAsync({
-      useClass: MariaDbConfigProvider,
-    }),
+    TypeOrmModule.forRoot(typeORMConfig),
     AdminOptionModule,
     ApplicationModule,
     BoardModule,
