+++ src/output/entities/BoardCategories.entity.ts
@@ -1,5 +1,5 @@
 import { Column, Entity, OneToMany, PrimaryGeneratedColumn } from 'typeorm';
-import { Boards } from './Boards';
+import { Boards } from './Boards.entity';
 
 @Entity('board_categories', { schema: 'dongurami_local_db' })
 export class BoardCategories {
